The principal file was moved to bin/
This is the original file with all tasks in one, for a more simple environment.  The "dividing tasks" files do the same job, but they have more adaptability capacities.  The main file was moved to a bin/ directory, from which it will be coppied to /usr/bin/ by the install script.  It also has just the task to call and coordinate the other scripts.